requirements.txt
PyQt6==6.6.1
SQLAlchemy==2.0.23
google-generativeai==0.3.2
Pillow>=10.2.0
python-dotenv==1.0.0
requests==2.31.0 